% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collin.r
\name{Collin}
\alias{Collin}
\title{Collin shiny app}
\usage{
Collin()
}
\value{
None
}
\description{
Launches the Collinearity shiny app, a tool to explore the consequences of collinearity on sampling distributions and inferential procedures in linear models
}
\details{
Full help is available from within the shiny app.
}
\examples{
## Not run:
if (interactive()){
 Collin()
 }
## End(Not run)
}
\author{
Francesco Pauli, \email{francesco.pauli@deams.units.it}
}
