% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfInt.r
\name{ConfInt}
\alias{ConfInt}
\title{ConfInt shiny app}
\usage{
ConfInt()
}
\value{
None
}
\description{
Launches the ConfInt shiny app, a tool to explore confidence intervals and regions for the coefficient of simple linear model
}
\details{
Full help is available from within the shiny app.
}
\examples{
## Not run:
if (interactive()){
 ConfInt()
 }
## End(Not run)
}
\author{
Francesco Pauli, \email{francesco.pauli@deams.units.it}
}
