% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessTheLine.r
\name{GuessTheLine}
\alias{GuessTheLine}
\title{GuessTheLine shiny app}
\usage{
GuessTheLine()
}
\value{
None
}
\description{
Launches the GuessTheLine shiny app, the user is prompted to guess the least squares fit of a sample of random points
}
\details{
Full help is available from within the shiny app.
}
\examples{
## Not run:
if (interactive()){
 GuessTheLine()
 }
## End(Not run)
}
\author{
Francesco Pauli, \email{francesco.pauli@deams.units.it}
}
