% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessThePoints.r
\name{GuessThePoints}
\alias{GuessThePoints}
\title{GuessThePoints shiny app}
\usage{
GuessThePoints()
}
\value{
None
}
\description{
Launches the GuessThePoints shiny app, the user is prompted to guess a sample of points compatible with a set of parameters.
}
\details{
Full help is available from within the shiny app.
}
\examples{
## Not run:
if (interactive()){
 GuessThePoints()
 }
## End(Not run)
}
\author{
Francesco Pauli, \email{francesco.pauli@deams.units.it}
}
