% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMBoard.R
\name{LMBoard}
\alias{LMBoard}
\title{LMBoard shiny app}
\usage{
LMBoard()
}
\value{
None
}
\description{
Launches the LMBoard shiny app, user can add remove points and visualize changes in the estimates and residual plots
}
\details{
Full help is available from within the shiny app.
}
\examples{
## Not run:
if (interactive()){
 LMBoard()
 }
## End(Not run)
}
\author{
Francesco Pauli, \email{francesco.pauli@deams.units.it}
}
