% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmviz-package.r
\docType{package}
\name{lmviz}
\alias{lmviz}
\alias{lmviz-package}
\title{lmviz: A package to visualize linear models features and play with them.}
\description{
The lmviz package contains three shyny apps
\describe{
  \item{\code{\link{SimpleLM}}}{allows to input the parameters of a linear model and simulate from it, exploring the various diagnostics and sampling distributions}
  \item{\code{\link{BadLM}}}{allows to set a data generating mechanism violating the homoscedasticity and/or normality assumptions of the linear model and explore by simulation its consequences on inference}
  \item{\code{\link{QuizResidual}}}{a game in which the user is prompted to guess whether the diagnostic plots of a linear model suggest some hypotheses is violated}
}
}
