% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleLM.R
\name{MultipleLM}
\alias{MultipleLM}
\title{MultipleLM shiny app}
\usage{
MultipleLM()
}
\value{
None
}
\description{
Launches the MultipleLM shiny app, a tool where the user can specify a multiple linearmodel, simulate from it and explore the sampling distribution of estimators under various conditions
}
\details{
Full help is available from within the shiny app.
}
\examples{
## Not run:
if (interactive()){
 MultipleLM()
 }
## End(Not run)
}
\author{
Francesco Pauli, \email{francesco.pauli@deams.units.it}
}
