% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.r
\name{Prediction}
\alias{Prediction}
\title{Prediction shiny app}
\usage{
Prediction()
}
\value{
None
}
\description{
Launches the Prediction shiny app, a tool to explore confidence and prediction intervals for the conditional mean in a simple linear model
}
\details{
Full help is available from within the shiny app.
}
\examples{
## Not run:
if (interactive()){
 Prediction()
 }
## End(Not run)
}
\author{
Francesco Pauli, \email{francesco.pauli@deams.units.it}
}
