% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDist.R
\name{SampleDist}
\alias{SampleDist}
\title{SampleDist shiny app}
\usage{
SampleDist()
}
\value{
None
}
\description{
Launches the SampleDist shiny app, a tool to explore sampling variability of the linear model and the sampling distributions of the least squares estimators
}
\details{
Full help is available from within the shiny app.
}
\examples{
## Not run:
if (interactive()){
 SampleDist()
 }
## End(Not run)
}
\author{
Francesco Pauli, \email{francesco.pauli@deams.units.it}
}
