% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleLM.R
\name{SimpleLM}
\alias{SimpleLM}
\title{SimpleLM shiny app}
\usage{
SimpleLM()
}
\description{
Launches the SimpleLM shiny app, which allows to input the parameters of a linear model and simulate from it, exploring the various diagnostics and sampling distributions
}
\details{
Full help is available from within the shiny app.
}
\examples{
## Not run:
if (interactive()){
 SimpleLM()
 }
## End(Not run)



}
\author{
Francesco Pauli, \email{francesco.pauli@deams.units.it}
}
