% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeTest.R
\name{SlopeTest}
\alias{SlopeTest}
\title{SlopeTest shiny app}
\usage{
SlopeTest()
}
\value{
None
}
\description{
Launches the SlopeTest shiny app, a tool to illustrate hypotheses testing on the slope in a simple regression model
}
\details{
Full help is available from within the shiny app.
}
\examples{
## Not run:
if (interactive()){
 SlopeTest()
 }
## End(Not run)
}
\author{
Francesco Pauli, \email{francesco.pauli@deams.units.it}
}
